fix build warning - trivial
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@16347 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_originate.c
@@ -1606,7 +1606,7 @@ static void *SWITCH_THREAD_FUNC early_thread_run(switch_thread_t *thread, void *
 					for (x = 0; x < (int)read_frame->datalen / 2; x++) {
 						sample = data[x] + mux_data[x];
 						switch_normalize_to_16bit(sample);
-						mux_data[x] = sample;
+						mux_data[x] = (int16_t)sample;
 					}
 					
 				}
